Remove unnecessary cast
git-svn-id: https://svn.apache.org/repos/asf/jakarta/jmeter/trunk@325300 13f79535-47bb-0310-9956-ffa450edef68

+++ src/core/org/apache/jmeter/reporters/FileReporter.java
@@ -411,8 +411,8 @@ class graphPanel extends JPanel
             g.setColor((Color) colorList.elementAt(t % colorList.size()));
             for (int i = 0; i < v.size() - 1; i++)
             {
-                float y1 = (float) ((Integer) v.elementAt(i)).intValue();
-                float y2 = (float) ((Integer) v.elementAt(i + 1)).intValue();
+                float y1 = ((Integer) v.elementAt(i)).intValue();
+                float y2 = ((Integer) v.elementAt(i + 1)).intValue();
 
                 y1 = y1 - minValue;
                 y2 = y2 - minValue;
